+++ src/CsvIterable.php
@@ -32,7 +32,7 @@ class CsvIterable implements CsvInterface
         $this->options = array_merge($this->options, $options);
     }
 
-    public function setMetadata(string $key, $value) 
+    public function setMetadata(string $key, $value)
     {
         $this->options[$key] = $value;
     }